requirements.txt
Django
djangorestframework
drf-yasg
python-dotenv
psycopg2-binary
